CursorAI/RouteAnalysys/requirements.txt
dash
pandas[parquet]
plotly
numpy
dash-bootstrap-components
pyyaml 